remove unused run_puzzle() function (it's only used by fast-forward tests and fuzzer)
